Constant URL UX
*Holding nose* Strategy here is to remove edit/show routes on games and
store state in the user to figure out where we are in the process
(current_game). Always redirect to users#show, which examines the state
of play and does what the old games controller did in edit/show, and
*renders* those views instead of redirecting. That means all the
partials have to be scoped. Still have the games update route, which
processes the game (but then redirects to user#show).

Also, removing the add bookmark button, though I guess it could stay if
we want it.